Fix build --install failure in SDK top level (#55)
When running build.py --install from top level, do not add clashing packaging/install rules.
The original if condition here only applied when --package was selected,
not for --install.
What we want is just to check if we are running from component level,
and only in that case bother with the packaging rules.

Change-Id: I79d943c742ea8d115d0f10afe953e5c1e7ce0459

Signed-off-by: Gergely Nagy <gergely.nagy@arm.com>